Enable argument configurations

+++ Yearly/make_lapse.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-while [[ $# > 1 ]]
+while [[ $# > 0 ]]
 do
     key="$1"
     case $key in
@@ -13,23 +13,24 @@ do
         -sm|--skip-make)
             SKIP_MAKE=1
             ;;
+        -sw|--skip-watermark)
+            SKIP_WATERMARK=1
+            ;;
         *)
             ;;
     esac
     shift # past argument or value
 done
 
-echo "$SKIP_COPY"
-echo "$SKIP_PREPARE"
-echo "$SKIP_MAKE"
-
-exit 0
-
 fps=30
 crf=18
 width=1296
 song=p2112104.mp4
-#convert=cp
+if [[ $SKIP_WATERMARK ]];
+then
+    echo "skip adding watermark to photos"
+    convert=cp
+fi
 
 rootdir=/root/fy16/
 dirsrc="$rootdir"selected/
@@ -44,26 +45,41 @@ arcfile="$dirdest"fy16-clip"$(date '+-%Y_%m_%d-%H_%M_%S')".mp4
 cd "$( dirname "${BASH_SOURCE[0]}" )"
 
 # copy photos
-rm -fr "$dirsrc"
-cp -v copy-yearly-photo.py "$rootdir"
-python3 "$rootdir"copy-yearly-photo.py
-
-# remove existing converted photos
-rm -fr "$dirtmp"
-mkdir "$dirtmp"
+if [[ $SKIP_COPY ]];
+then
+    echo "skip copying photos"
+else
+    rm -fr "$dirsrc"
+    cp -v copy-yearly-photo.py "$rootdir"
+    python3 "$rootdir"copy-yearly-photo.py
+fi
 
 # prepare photos
-count=0
-for filesrc in $(ls "$dirsrc"*.jpg)
-do
-    count=`expr $count + 1`
-    filedest=`printf ""$dirtmp"snap_%.4d.jpg" "$count"`
-    ./convert_frame.sh "$filesrc" "$filedest" "$convert"
-done
+if [[ $SKIP_PREPARE ]];
+then
+    echo "skip preparing photos"
+else
+    rm -fr "$dirtmp"
+    mkdir "$dirtmp"
 
-# generate h264 version video for modern browsers' users
-ffmpeg -y -vn -f image2 -i "$dirtmp"snap_%04d.jpg -i "$bgmmp4" -shortest -r "$fps" -vf scale="$width":-1 -vcodec libx264 -crf "$crf" -tune stillimage -profile:v high -level 4.2 -acodec aac -ab 128k -strict experimental "$filemp4"
+    count=0
+    for filesrc in $(ls "$dirsrc"*.jpg)
+    do
+        count=`expr $count + 1`
+        filedest=`printf ""$dirtmp"snap_%.4d.jpg" "$count"`
+        ./convert_frame.sh "$filesrc" "$filedest" "$convert"
+    done
+fi
 
-# archive & share it
-mv -v "$filemp4" "$arcfile"
-cp -v "$arcfile" "$dirpublic"
+# create video and upload
+if [[ $SKIP_MAKE ]];
+then
+    echo "skip making video"
+else
+    # generate h264 version video for modern browsers' users
+    ffmpeg -y -vn -f image2 -i "$dirtmp"snap_%04d.jpg -i "$bgmmp4" -shortest -r "$fps" -vf scale="$width":-1 -vcodec libx264 -crf "$crf" -tune stillimage -profile:v high -level 4.2 -acodec aac -ab 128k -strict experimental "$filemp4"
+
+    # archive & share it
+    mv -v "$filemp4" "$arcfile"
+    cp -v "$arcfile" "$dirpublic"
+fi
